[DEV-3385] open file at root domain
